client: Use x86-64 indexer for animation number

+++ game/client/game_controls/basemodelpanel.h
@@ -227,7 +227,7 @@ public:
 
 private:
 	bool	m_bPanelDirty;
-	int		m_iDefaultAnimation;
+	intp	m_iDefaultAnimation;
 
 	bool	m_bAllowOffscreen;
